pyproject.toml
[tool.poetry]
name = "broutonlab-python-template"
version = "0.1.0"
description = ""
authors = ["Yuriy Turchin <yturchin17@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pytorch-lightning = "^1.6.1"
torchvision = "^0.12.0"
hydra-core = "^1.1.2"
albumentations = "^1.1.0"
pandas = "^1.4.2"
mlflow = "^1.25.1"
black = "21.12b0"
click = "8.0.4"
matplotlib = "^3.5.2"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
mypy = "^0.930"
flake8 = "^4.0.1"
flake8-bugbear = "^21.11.29"
flake8-builtins = "^1.5.3"
flake8-comprehensions = "^3.7.0"
flake8-docstrings = "^1.6.0"
flake8-import-order = "^0.18.1"
flake8-tidy-imports = "^4.5.0"
pytest = "^6.2.5"
distlib = "0.3.3"
pre-commit = "^2.17.0"

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
force_sort_within_sections = true
line_length = 79
order_by_type = false
known_first_party=["app", "src", "legacy_pg"]

[tool.black]
line-length = 79
exclude = "(^cdk/| ^point_generation/tests/)"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
